Fixes test_absolute_limits testing the wrong action.
Absolute Limits is currently deprecated (this can be confirmed
by executing nova help absolute-limits). So the file name should
be changed to be test_limits_rbac.

In addition, the action os_compute_api:os-used-limits is "tested",
but in reality this is not the case: the policy is not actually
enforced by the test. The policy action should be changed to
os_compute_api:limits instead.

This patch set changes the following:
  - Corrects file name and class name
  - Removes unnecessary clients being delcared in setup_clients
  - Corrects the action being tested

Change-Id: I97a0015fab1ac1607c4d0ef07180ca98428c5fb2