Changes tox to only run unit tests and moves unit tests to tests/unit.
Currently, the api and scenario tests are skipped when tox is run,
i.e. when tox -e py27 and tox -e py35 are executed. This is not ideal,
because an error could be thrown when trying to import modules into
the api and scenario tests. Instead, these tests should never
be discovered by ostestr. This will resolve issues with the new
framework changes [0].

This change simply creates a new file called test-whitelist.txt
which is used by the parameter --whitelist-file in the
command under [testenv] in tox.ini. It ensures that only
the regexes specified in test-whitelist.txt are executed [1] --
so only the unit tests are executed with tox (as currently
no functional tests are in Patrole).

Also added check-uuid to pep8 in tox.ini.

[0] https://review.openstack.org/#/c/433741/
[1] https://docs.openstack.org/developer/os-testr/ostestr.html

Change-Id: I1c6bca9044a74e8909754f106880636739e5a822