Add negative lookahead to post_test_hook to skip slow tests.
This will skip a few slow tests in compute and volume with the
@test.attr(type='slow') decorator above that, so that the patrole
tests that run in the tempest gates can finish in a reasonable
amount of time.

The post_test_hook can later be modified, once new gates are
introduced in infra, to run the slow tests independent from
the "fast" tests.

Change-Id: I794a5da46a11bedd92a91d4b4c1f69f3d6e4e804
Partially-Implements: blueprint decrease-tempest-gate-runtime