Switch to admin role during client set up to fix some gate bugs.
The current framework only adds the admin role to
the cls.os.<client-name> namespace when switchToRbacRole=False
is executed; meaning that the first test that is executed
in a test run will not have the admin role during set up.

Thus, it is necessary to add a switch_role with
switchToRbacRole=False to setup_clients (after rbac_utils
is instantiated), so that class-level resources are set
up with admin credentials. This is why
test_flavor_extra_specs_rbac is currently failing [0].
(setup_clients is called before resource_setup.)

This patch:
  * Performs switchToRbacRole=False in the setup_clients
    of every base Patrole test class.

[0] http://logs.openstack.org/94/444994/2/check/gate-tempest-dsvm-patrole-admin-ubuntu-xenial-nv/a00e22a/console.html

Change-Id: I8027938daa4fdfff0ac5336324aa6d647322237b
Partial-Bug: #1670553