Update Cinder test that incorrectly handles 404
The volume test that checks the "volume:delete" action incorrectly
handles a 404. When a role doesn't have permission to delete a
volume, Cinder actually throws a 403 and not a 404.

This commit also refactors existing test to group basic volume CRUD
(create, read, update, and delete) tests into a single file.

Additionally, there was not a test for "volume:update", so this commit
also covers that action.

Change-Id: Ib5f63ecc1ac51918ee4b1e35763b0c91aa7d5ae2
Closes-Bug: #1671903