Fixes failing flavor access tests for Member role.
Fixes the test for policy action os_compute_api:os-flavor-access,
which currently throws a 404 as Member, even though the policy
action is defined as admin_or_owner. This means that Member should
be allowed to pass the policy rule, but previously the API call
flavors_client.list_flavor_accesss was used which calls other
policy actions which disallow Member to do the API call.

This patch uses a different API--flavors_client.show_flavor--which
calls the relevant policy action. The test passes, so long
as a public flavor is used. Since CONF.compute.flavor_ref is by
default public, it is used instead of dynamically creating a
public flavor, which adds unnecessary overhead.

Change-Id: I107c4daa48cf6922cee1746dc1fc8490b542fc8a
Partial-Bug: #1670553