Partially revert removal of time.sleep if v3 auth enabled in conf.
In a previous patch [0], an attempt was made to speed up Patrole
tests by skipping over time.sleep(1) if v3 auth was enabled
in tempest.conf.

However, this resulted in more tests failing than otherwise should
be for non-admin roles (i.e. Member role). This is because:

"Fernet tokens are not subsecond aware and Keystone should only be
 precise to the second. Sleep to ensure we are passing the second
 boundary before attempting to authenticate." [1][2]

This patch guarantees that time.sleep(1) is always executed,
regardless of identity auth version used by Tempest. While this
prolongs the test runs in the gates, this change is necessary
to guarantee correct test execution for non-admin roles.

[0] https://review.openstack.org/#/c/444967/
[1] https://github.com/openstack/tempest/blob/master/tempest/api/identity/v2/test_users.py
[2] https://github.com/openstack/tempest/blob/master/tempest/api/identity/v3/test_users.py

Change-Id: I007190906e710f540ed0e44e76bd17047fcddea4
Closes-Bug: #1673497