Update post_test_hook to use multinode environment.
Now that a new infra job has been added to Patrole for a
multinode environment [0], the post_test_hook in Patrole
needs to be updated.

This change will still keep "fast" tests in the "fast" gates,
although they have been renamed to not have the "fast" substring;
and it will still keep "slow" tests in the "slow" gates,
although they have been renamed with "multinode" instead of "slow".

This means that tests that require multiple nodes, like migrating
a server, can be run, rather than skipped.

[0] https://review.openstack.org/#/c/450278/

Change-Id: I278695026ce273daa8cb44e9c472cf84e880a618