Renames switchToRbacRole to toggle_rbac_role
This is a trivial change that renames switchToRbacRole to
toggle_rbac_role.

This change is needed because:
  - switchToRbacRole is camel case which is inconsistent with
    all the snake-case variables names throughout Patrole and
    OpenStack
  - It violates Python's naming conventions [0].

[0] See pep8 styling guideline:

"Function names should be lowercase, with words separated
by underscores as necessary to improve readability. mixedCase
is allowed only in contexts where that's already the prevailing
style."

And for variable names:

"Use the function naming rules: lowercase with words separated
by underscores as necessary to improve readability."

Change-Id: Ib6dd50ba02c25d585e2ca0c9383771f9337212b8
Depends-On: Ie1c4b6a2801f10c8e83f6d3f88fe0364d69317fa