Fix compute create volume test race condition
Currently, create_volume test in test_volume_rbac in compute
module can fail [0] because of a race condition associated
with not waiting for the volume to reach 'active' status
following creation. Normally this is not an issue because
the class helper method "create_volume" is used which
does this in the background, but in this particular test,
create_volume is called via the volumes_extensions_client,
so no wait is currently being done.

This patch adds a wait for the volume to reach 'active'
status.

[0] http://logs.openstack.org/71/466871/4/check/gate-tempest-dsvm-patrole-admin-ubuntu-xenial/fbb5290/console.html

Change-Id: I8c16fab69502fe94e39a63afbb94eee72cc4cac1