[Gate fix] Fix test_server_actions raising BadRequest
When validatable is set to True for
create_test_server, a BadRequest is thrown, with
details that 'Instance network is not ready yet.'

A workaround is to remove validatable=True from
calls to create_test_server, because it is not
necessary for Patrole testing. The docs
for create_test_server are:

:param validatable: Whether the server will be
                    pingable or sshable.

So we can remove it.

Change-Id: I5e963cd2b8649ed8ebe35cf776432bd2496bdd27