Fix volume backup tests throwing BadRequest
Currently, if a volume fail to backup because of a timeout
exception, then the volume may be in backing-up state for the
next test, causing that test to in turn fail [0].

This commit prevents cascading failures across tests by
creating a volume during each test run.

This commit also replaces ``_create_backup`` with ``create_backup``
in Tempest's base volume class to simplify code.

[0] http://logs.openstack.org/58/469258/1/check/gate-tempest-dsvm-patrole-multinode-admin-ubuntu-xenial-nv/03de6d3$

Change-Id: Iba788e1947087a6ffa59137087ac00087507f26a