RBAC test for update_subnetpool:is_default
This commit adds a new RBAC test for subnetpools for updating
a default subnetpool. Because the update may be performed on a
pre-existing default subnetpool (because only 1 default subnetpool
may exist per IP family -- 4 and 6), the update operation only
updates the description to the already-existing description, in
effect doing nothing to the subnetpool but still doing policy
enforcement for the specified policy action.

In addition, this commit does not add a test for
create_subnetpool:is_default because in all likelihood default
subnetpools already exist and it is not safe to delete them and
re-create them for multiple reasons.

Change-Id: Ic2f5436f80354c76d2dbd404e3966e34088e86f3