Fix test_volumes_snapshots_rbac throwing BadRequest
test_volumes_snapshots_rbac frequently throws a BadRequest
during cleanup with the following message: Invalid snapshot:
Snapshot status must be available or error and must not be part of
a group.

This fix adds waiters to certain snapshot operations so that
cleanup is not attempted while this snapshots are in a state
that would cause this BadRequest error to be thrown.

Change-Id: Ia09f692c8597c9e776dfeca4e67e970d54965cda
Closes-Bug: #1694500