docs: Update configuration docs
This commit updates configuration docs. It also refactors the
index doc page to make it look more like Tempest's. The "readme"
file was removed from the docs as it's just a high-level project
overview.

The Developer's Guide section will be updated with contents included
in the "readme" in far greater detail in a follow-up commit.

Change-Id: I4d468b7f1bb6a000fde42d656635159176d5ef7f