Switch to enabled version of identity clients
Patrole always uses (e.g.) v3 roles client to retrieve
list of roles which is bad if the v3 identity service is not enabled.

Cases like the following:

    self.roles_client = test_obj.os_admin.roles_v3_client

Should be changed to:

    self.roles_client = test_obj.os_admin.roles_v3_client \
                        if CONF.identity_feature_enabled.api_v3 \
                        else test_obj.os_admin.roles_client

This commit switches between the correct identity client
depending on the identity version that is enabled in tempest.conf.
The v3 client is prioritized as identity v3 is current.

This commit also corrects/improves upon some documentation errata.

Change-Id: I9a12196f11473ac4e045ae90c4321219beab7ca6
Closes-Bug: #1702980