Change rbac_utils.RbacUtils is_admin to function
This is a trivial change that changes is_admin to a function
(instead of a property) because properties can only be used
in the instance context whereas non-class functions can be used
in both instance and class contexts.

This allows ``test_tokens_negative_rbac`` to use is_admin()
now since it is used in a class context.

Change-Id: I9244f10cd536050986185651421a446fef8e2ade