Use configured admin creds in rbac utils
Use ``get_configured_admin_credentials()`` in rbac_utils to instantiate
an admin client manager without having to rely on creating a new
set of credentials, which is problematic when using pre-provisioned
credentials. This is because rbac_utils consumes a set of credentials
allocated in the accounts.yaml file, which can cause tests to fail.

Using ``get_configured_admin_credentials()`` is a workaround because
it uses admin credentials specified in tempest's config's [auth]
section, which already exist in the system.

Change-Id: I0622bb057c365fdf652e4a110125678371ac4569