Add missing v3 volume tests for which v2 tests exist
This commit adds missing v3 volume tests for which corresponding
v2 tests exist, and if the v2 API is also present in v3.

This commit is necessary for partially deprecating the v2 volume
RBAC tests -- but before that can be done, we need to ensure
that v3 has at least the exact same coverage v2 does, plus
additional coverage for v3 APIs.

This commit specifically adds v3 tests for:
  - test_qos_rbac
  - test_volume_basic_crud_rbac
  - test_volume_hosts_rbac
  - test_volume_metadata_rbac

This commit also adds a missing RBAC test for showing a volume
host for both v2 and v3.

Depends-On: I09ad8200f73273dae8e5d926939620b71d9210e8
Change-Id: I443dfb8e1032ff327f03eb8e8b257df124ad6180
Partially Implements: blueprint test-v3-volume-api-only