Test coverage for compute flavor_manage policies
This commit adds test coverage for the flavor_manage policies,
which includes tests only for the non-deprecated policy actions:

* "os_compute_api:os-flavor-manage:create"
* "os_compute_api:os-flavor-manage:delete"

This commit also renames _create_flavor in rbac_base for compute
module to be public since it is consumed by other files. It also
adds a wait_for_resource_deletion call.

Change-Id: I632235a1d81488ecc8fb1e92f801415a08c7e778