Remove identity v3 change_password test
This commit removes the v3 change_password test because
it is currently getting skipped [0], because the policy
action does not exist in [1] and the API where it would
be enforced does not enforce it [2].

So the correct approach is to remove the test. Also renames
the class name to drop "Admin" substring as there is nothing
about this test class that requires admin credentials in any
way.

[0] http://logs.openstack.org/68/496068/4/check/gate-tempest-dsvm-patrole-admin-ubuntu-xenial/7069c22/console.html#_2017-08-23_00_50_18_947787
[1] https://github.com/openstack/keystone/blob/master/keystone/common/policies/user.py
[2] https://github.com/openstack/keystone/blob/4950c24bddaf98b79dfd9ee178ae1086cf48e322/keystone/identity/controllers.py#L279

Change-Id: I90496d3969b1ef69316a2254754622679bca54b5