Fix router tests expecting wrong error code
Currently tests in test_router_rbac expect a 404 for some endpoints
that actually raise a 403 instead. The reason why the tests currently
pass in the gates for admin and Member is because the associated
policies are admin_or_owner -- meaning that the 403 or 404 is never
raised, causing the tests to silently pass. But if elevated permissions
are used then a 403 -- not a 404 -- is raised.

The except is for the test "test_show_router" which enforces `get_router`
which does raise a 404, so that is left unchanged.

Change-Id: If37e0b15137d1421f9d16fc515068628ab87768f